Backend/requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-jose
passlib
python-multipart
pdfplumber
httpx
python-dotenv
pydantic-settings
